requirements.txt
python-pptx>=0.6.21
pdf2image>=1.16.0
Pillow>=9.0.0